Extract PrimitiveSerialization
